[artCharcoal] Avoid giving global scope for the main processing function
